fixes for 64-bit and boost 1.32
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@18037 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ loreley/src/libloreley/md5.cc
@@ -203,7 +203,7 @@ md5::_md5init(md5::ctx_t *mdContext)
 }
 
 void
-md5::_md5update (md5::ctx_t *mdContext, unsigned char const *inBuf, unsigned int inLen)
+md5::_md5update (md5::ctx_t *mdContext, unsigned char const *inBuf, size_t inLen)
 {
   uint32_t in[16];
   int mdi;
